requirements-minimal.txt
fastapi
uvicorn
sqlalchemy
google-api-python-client
python-dotenv
scikit-learn
pandas
numpy
joblib
